backend/requirements.txt
# ── Core Framework ──────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Database ────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# ── Cache / Pub-Sub ────────────────────────────
redis==5.2.1

# ── Security ────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
email-validator==2.1.1

# ── WebSocket ───────────────────────────────────
websockets==14.1

# ── Docker SDK ──────────────────────────────────
docker>=7.0

# ── Utilities ───────────────────────────────────
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1

# ── ML / AI (installed separately — large packages) ──
# torch, numpy, scikit-learn are installed via
# requirements-ml.txt when needed for inference
